Move context menu ordering function to shared
