+++ packages/shared/src/ContextMenuPresenter.js
@@ -1,5 +1,3 @@
-// Provide a context, get context widgets back
-
 const PUBLIC_CONTEXT_WIDGETS = [
   { context: 'public', title: 'widget-public-stream', id: 'widget-public-stream', view: 'stream', order: 1, parentId: null },
   { context: 'public', title: 'widget-public-groups', id: 'widget-public-groups', view: 'groups', order: 2, parentId: null },
@@ -41,4 +39,35 @@ export function getStaticMenuWidgets ({ isPublic, isMyContext, profileUrl, isAll
   }
 
   return widgets
+}
+
+export const orderContextWidgetsForContextMenu = (contextWidgets) => {
+  // Step 1: Filter out widgets without an order, as these are not displayed in the context menu
+  const orderedWidgets = contextWidgets.filter(widget => widget.order !== null)
+
+  // Step 2: Split into parentWidgets and childWidgets
+  const parentWidgets = orderedWidgets.filter(widget => !widget?.parentId)
+  const childWidgets = orderedWidgets.filter(widget => widget?.parentId)
+
+  // Step 3: Add an empty array for childWidgets to each parentWidget
+  parentWidgets.forEach(parent => {
+    parent.childWidgets = []
+  })
+
+  // Step 4: Append each childWidget to the appropriate parentWidget
+  childWidgets.forEach(child => {
+    const parent = parentWidgets.find(parent => parent.id === child?.parentId)
+    if (parent) {
+      parent.childWidgets.push(child)
+    }
+  })
+
+  // Step 5: Sort parentWidgets and each childWidgets array by order
+  parentWidgets.sort((a, b) => a.order - b.order)
+  parentWidgets.forEach(parent => {
+    parent.childWidgets.sort((a, b) => a.order - b.order)
+  })
+
+  // Return the sorted parentWidgets with nested childWidgets
+  return parentWidgets
 }
